[SCISPARK 21] Refactor GTG
